+++ src/app/settings.rs
@@ -1,9 +1,12 @@
+use crate::app::fonts::FontSizes;
+
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::path::{Path, PathBuf};
 
 pub const FILENAME: &str = "dockeye.yml";
+const ALLOWED_FONT_SIZE: std::ops::RangeInclusive<f32> = 10.0..=50.0;
 
 pub fn dir() -> Option<PathBuf> {
     dirs::config_dir()
@@ -12,12 +15,14 @@ pub fn dir() -> Option<PathBuf> {
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Settings {
     pub docker_addr: String,
+    pub fonts: FontSizes,
 }
 
 impl Default for Settings {
     fn default() -> Self {
         Self {
             docker_addr: crate::DEFAULT_DOCKER_ADDR.to_string(),
+            fonts: FontSizes::default(),
         }
     }
 }
@@ -100,6 +105,10 @@ impl SettingsWindow {
 "#,
                         );
                     ui.end_row();
+
+                    self.fonts_ui(ui);
+                    ui.end_row();
+
                     if ui.button("save").clicked() {
                         if let Err(e) = self.save_settings() {
                             msg = Some(Message::Error(format!("{:?}", e)));
@@ -117,5 +126,43 @@ impl SettingsWindow {
             });
         self.show = show;
         self.msg = msg;
+    }
+    fn fonts_ui(&mut self, ui: &mut egui::Ui) {
+        egui::CollapsingHeader::new("fonts")
+            .default_open(false)
+            .show(ui, |ui| {
+                egui::Grid::new("fonts_grid").show(ui, |ui| {
+                    ui.label("small");
+                    ui.add(
+                        egui::DragValue::new(&mut self.settings.fonts.small)
+                            .clamp_range(ALLOWED_FONT_SIZE),
+                    );
+                    ui.end_row();
+                    ui.label("body");
+                    ui.add(
+                        egui::DragValue::new(&mut self.settings.fonts.body)
+                            .clamp_range(ALLOWED_FONT_SIZE),
+                    );
+                    ui.end_row();
+                    ui.label("button");
+                    ui.add(
+                        egui::DragValue::new(&mut self.settings.fonts.button)
+                            .clamp_range(ALLOWED_FONT_SIZE),
+                    );
+                    ui.end_row();
+                    ui.label("heading");
+                    ui.add(
+                        egui::DragValue::new(&mut self.settings.fonts.heading)
+                            .clamp_range(ALLOWED_FONT_SIZE),
+                    );
+                    ui.end_row();
+                    ui.label("monospace");
+                    ui.add(
+                        egui::DragValue::new(&mut self.settings.fonts.monospace)
+                            .clamp_range(ALLOWED_FONT_SIZE),
+                    );
+                    ui.end_row();
+                });
+            });
     }
 }