+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add ability to change docker socket location in settings
 - Display a confirmation popup when deleting a container/image
 - A lot of ui fixes and tweaks, more readable text
+- Add configurable font sizes to the settings window